requirements.txt
amqp==5.0.6
asgiref==3.3.4
attrs==21.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.0.5
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.2.2
django-crispy-forms==1.12.0
django-debug-toolbar==3.2
django-extensions==3.1.2
djangorestframework==3.12.4
Faker==8.0.0
flake8==3.9.0
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
kombu==5.0.2
mccabe==0.6.1
packaging==20.9
Pillow==8.3.0
pluggy==0.13.1
prompt-toolkit==3.0.18
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
python-dateutil==2.8.1
python3-memcached==1.51
pytz==2021.1
rabbitmq==0.2.0
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.6
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
XlsxWriter==1.4.2
